+++ vecorel_cli/registry.py
@@ -1,5 +1,6 @@
 import importlib.metadata
 import re
+from typing import Union
 
 
 class VecorelRegistry:
@@ -32,8 +33,11 @@ class VecorelRegistry:
         "collection",
     ]
 
-    # Any schemas that are mandatory to have, except for the vecorel specification schema
-    required_extensions: list[str, re.Pattern] = []
+    # Any schemas patterns that are mandatory to have, except for the vecorel specification schema
+    required_extensions: list[Union[str, re.Pattern]] = []
+
+    # Any specific schemas URIs that are added by default when creating a new dataset
+    default_extensions: list[str] = []
 
     # The filenames for datasets (converters) that should be ignored by the CLI.
     # Always ignores files with a starting "." or "__"
@@ -78,6 +82,8 @@ class VecorelRegistry:
 
         if isinstance(command, BaseCommand):
             command_name = command.cmd_name
+        else:
+            command_name = command
         self.commands.pop(command_name, None)
 
     def set_command(self, command):